refactored to remove nuisance utils and refactor to registration utils
